+++ src/lib/exceptions.js
@@ -1,5 +1,6 @@
 export default Object.freeze({
     INVALID_API_TOKEN: 'Invalid API Token',
     INSUFFICIENT_PERMISSIONS: 'Insufficient permissions for given API token',
-    METHOD_NOT_IMPLEMENTED: 'Method not implemented'
+    METHOD_NOT_IMPLEMENTED: 'Method not implemented',
+    INVALID_SQL_QUERY: 'Invalid SQL query'
 });