+++ src/modules/query.js
@@ -1,20 +1,28 @@
 import { getLogger } from '../lib/logger';
 const logger = getLogger('query-module');
 import { fetch } from '../lib/http';
-import { sanitizeSQL } from '../lib/utils';
+import { sanitizeSQL, isValidSQLQuery } from '../lib/utils';
+import Exceptions from '../lib/exceptions';
 
 module.exports = {
     /**
      * Execute SQL query
      * 
-     * @param {string} sql SQL Query
+     * @param  { String } sql SQL Query
+     * @return { Object } resultset rows
      */
     query: async sql => {
+        let sanitizedQuery;
         try {
-            return fetch(`/v0/sql?q=${sanitizeSQL(sql)}`);
+            if (!isValidSQLQuery(sql)) {
+                throw new Error(Exceptions.INVALID_SQL_QUERY);
+            }
+
+            sanitizedQuery = sanitizeSQL(sql);
+            return fetch(`/v0/sql?q=${sanitizedQuery}`);
         } catch (error) {
             logger.error('Error while fetching /v0/sql');
-            logger.debug(`Query: ${sanitizeSQL}`);
+            logger.debug(`Query: ${sanitizedQuery}`);
             logger.debug(error);
         }
     }