+++ CHANGELOG.md
@@ -6,6 +6,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+- Add `query` method
+- Validate and sanitize SQL queries
+- More tests
+- Update docs
+- Update package name
+
 ## [0.1.0] - 2022-05-31
 
 - Create basic project structure